+++ Assets/Features/HealthView.cs
@@ -5,7 +5,7 @@ public class HealthView : MonoBehaviour, IHealthView
 {
     [SerializeField] private Slider _healthSlider;
     
-    public void UpdateHealth(int currentHealth, int maximumHealth)
+    public void UpdateHealth(float currentHealth, float maximumHealth)
     {
         _healthSlider.maxValue = maximumHealth;
         _healthSlider.value = currentHealth;
